Assignments/Quarter-4/LangChain-RAG_Basic/pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = ""
authors = ["Abrar-Kamoka <abrarabbas.ak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.7"
langchain = "^0.3.13"
langchain-text-splitters = "^0.3.4"
langchain-community = "^0.3.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
